refactor: change scripts structure
