+++ scripts/main.js
@@ -1,25 +1,9 @@
 'use strict';
 
 import { keyboard } from './keyboardObject.js';
+import { defineButton, changeCharactersToUppercase, changeCharactersToLowercase } from './eventHandlers.js';
 
 keyboard.init();
-
-function defineButton(event) {
-    let buttonSelector;
-    let keyCode = String(event.keyCode);
-    if (keyCode == '16' || keyCode == '17' || keyCode == '18') {
-        if (event.location == 1) {
-            buttonSelector = `button[keyCode='${keyCode}'][side='left']`;
-        }
-        if (event.location == 2) {
-            buttonSelector = `button[keyCode='${keyCode}'][side='right']`;
-        }
-    } else {
-        buttonSelector = `button[keyCode='${keyCode}']`;
-    }
-    let button = document.querySelector(buttonSelector);
-    return button;
-}
 
 window.addEventListener('keydown', (event) => {
     let keyCode = String(event.keyCode);
@@ -31,34 +15,10 @@ window.addEventListener('keydown', (event) => {
             } else {
                 button.removeAttribute('active', '');
             }
-            let characterButtons = document.querySelectorAll('.character');
-            if (keyboard.currentLang == 'RU') {
-                characterButtons.forEach(characterButton => {
-                    let keyCode = characterButton.getAttribute('keycode');
-                    characterButton.textContent = keyboard.buttonsList[keyCode].defValueUp;
-                });
-            }
-            if (keyboard.currentLang == 'EN') {
-                characterButtons.forEach(characterButton => {
-                    let keyCode = characterButton.getAttribute('keycode');
-                    characterButton.textContent = keyboard.buttonsList[keyCode].altValueUp;
-                });
-            }
+            changeCharactersToUppercase();
         }
         if (button.getAttribute('keycode') == '16') {
-            let characterButtons = document.querySelectorAll('.character');
-            if (keyboard.currentLang == 'RU') {
-                characterButtons.forEach(characterButton => {
-                    let keyCode = characterButton.getAttribute('keycode');
-                    characterButton.textContent = keyboard.buttonsList[keyCode].defValueUp;
-                });
-            }
-            if (keyboard.currentLang == 'EN') {
-                characterButtons.forEach(characterButton => {
-                    let keyCode = characterButton.getAttribute('keycode');
-                    characterButton.textContent = keyboard.buttonsList[keyCode].altValueUp;
-                });
-            }
+            changeCharactersToUppercase();
         }
         button.classList.add('pressed');
     }
@@ -70,36 +30,12 @@ window.addEventListener('keyup', (event) => {
         let button = defineButton(event);
         if (button.getAttribute('keycode') == '20') {
             if (!button.hasAttribute('active')) {
-                let characterButtons = document.querySelectorAll('.character');
-                if (keyboard.currentLang == 'RU') {
-                    characterButtons.forEach(characterButton => {
-                        let keyCode = characterButton.getAttribute('keycode');
-                        characterButton.textContent = keyboard.buttonsList[keyCode].defValue;
-                    });
-                }
-                if (keyboard.currentLang == 'EN') {
-                    characterButtons.forEach(characterButton => {
-                        let keyCode = characterButton.getAttribute('keycode');
-                        characterButton.textContent = keyboard.buttonsList[keyCode].altValue;
-                    });
-                }
+                changeCharactersToLowercase();
                 button.classList.remove('pressed');
             }
         } else {
             if (button.getAttribute('keycode') == '16') {
-                let characterButtons = document.querySelectorAll('.character');
-                if (keyboard.currentLang == 'RU') {
-                    characterButtons.forEach(characterButton => {
-                        let keyCode = characterButton.getAttribute('keycode');
-                        characterButton.textContent = keyboard.buttonsList[keyCode].defValue;
-                    });
-                }
-                if (keyboard.currentLang == 'EN') {
-                    characterButtons.forEach(characterButton => {
-                        let keyCode = characterButton.getAttribute('keycode');
-                        characterButton.textContent = keyboard.buttonsList[keyCode].altValue;
-                    });
-                }
+                changeCharactersToLowercase();
             }
             button.classList.remove('pressed');
         }
